smaa_areatex: Improve handling of diagonal U patterns, don't straighten them
The older method that straightens diagonal U patterns yielded ugly distorted
circles. This implementation uses a triangle to calculate the occupied areas.
Note that this was tested only for SMAA 1x, not for the other modes doing
subpixel rendering.

We still can use the older one by a command line option "-u".